Factor out loadable module handling from PluginManager.
Behavior is mostly unchanged, with one major exception: when a module is
requested, the previously loaded version (if any) is always used, instead of
searching the paths anew. This prevents different files from being loaded for
the same module name (unless the module is first unloaded) in the case where
the search paths or the files within them have changed. (It probably also
speeds up device loading.)

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12462 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd